Trying to convert disagg_by_rel_sources into a postprocessor
