+++ openquake/calculators/classical.py
@@ -18,7 +18,6 @@
 
 import io
 import os
-import re
 import time
 import psutil
 import logging
@@ -633,8 +632,6 @@ class ClassicalCalculator(base.HazardCalculator):
                 raise RuntimeError('%s in #%d' % (msg, self.datastore.calc_id))
             elif slow_tasks:
                 logging.info(msg)
-        if 'rates_by_src' in self.datastore and self.N == 1 and len(oq.poes):
-            store_mean_disagg_bysrc(self.datastore, self.csm)
 
     def _create_hcurves_maps(self):
         oq = self.oqparam
@@ -778,75 +775,3 @@ class ClassicalCalculator(base.HazardCalculator):
             smap = parallel.Starmap(make_hmap_png, allargs)
             for dic in smap:
                 self.datastore['png/hmap_%(m)d_%(p)d' % dic] = dic['img']
-
-# ######################### postprocessing ################################### #
-
-
-def get_rel_source_ids(dstore, imts, poes, threshold=.1):
-    """
-    :returns: sorted list of relevant source IDs
-    """
-    source_ids = set()
-    for im in imts:
-        for poe in poes:
-            aw = extract.extract(dstore, f'rates_by_src?imt={im}&poe={poe}')
-            poe_array = aw.array['poe']  # for each source in decreasing order
-            max_poe = poe_array[0]
-            rel = aw.array[poe_array > threshold * max_poe]
-            source_ids.update(rel['src_id'])
-    return python3compat.decode(sorted(source_ids))
-
-
-def middle(arr):
-    """
-    :returns: middle values of an array (length N -> N-1)
-    """
-    return [(m1 + m2) / 2 for m1, m2 in zip(arr, arr[1:])]
-
-
-def store_mean_disagg_bysrc(dstore, csm):
-    """
-    Compute and store the mean disaggregatiob by Mag_Dist_Eps for
-    each relevant source in the source model
-    """
-    parent = dstore.parent or dstore
-    oq = parent['oqparam']
-    # oq.cachedir = datastore.get_datadir()
-    oq.mags_by_trt = {
-                trt: python3compat.decode(dset[:])
-                for trt, dset in parent['source_mags'].items()}
-    sitecol = parent['sitecol']
-    assert len(sitecol) == 1, sitecol
-    edges, shp = disagg.get_edges_shapedic(oq, sitecol)
-    if 'rates_by_src' in parent:
-        rel_ids = get_rel_source_ids(parent, oq.imtls, oq.poes, threshold=.1)
-    else:
-        rel_ids = get_rel_source_ids(dstore, oq.imtls, oq.poes, threshold=.1)
-    logging.info('There are %d relevant sources: %s',
-                 len(rel_ids), ' '.join(rel_ids))
-
-    smap = parallel.Starmap(disagg.disagg_source, h5=dstore.hdf5)
-    src2idx = {}
-    for idx, source_id in enumerate(rel_ids):
-        src2idx[source_id] = idx
-        smlt = csm.full_lt.source_model_lt.reduce(source_id, num_samples=0)
-        gslt = csm.full_lt.gsim_lt.reduce(smlt.tectonic_region_types)
-        relt = FullLogicTree(smlt, gslt)
-        Z = relt.get_num_paths()
-        assert Z
-        logging.info('Considering source %s (%d realizations)',
-                     source_id, Z)
-        groups = relt.reduce_groups(csm.src_groups, source_id)
-        assert groups, 'No groups for %s' % source_id
-        smap.submit((groups, sitecol, relt, (edges, shp), oq))
-    mags, dists, lons, lats, eps, trts = edges
-    arr = numpy.zeros(
-        (len(rel_ids), shp['mag'], shp['dist'], shp['eps'], shp['M'], shp['P']))
-    for srcid, rates5D, rates2D in smap:
-        idx = src2idx[basename(srcid, '!;')]
-        arr[idx] = disagg.to_probs(rates5D)
-    dic = dict(
-        shape_descr=['source_id', 'mag', 'dist', 'eps', 'imt', 'poe'],
-        source_id=rel_ids, imt=list(oq.imtls), poe=oq.poes,
-        mag=middle(mags), dist=middle(dists), eps=middle(eps))
-    dstore['mean_disagg_bysrc'] = hdf5.ArrayWrapper(arr, dic)
