+++ openquake/qa_tests_data/logictree/case_04/job.ini
@@ -3,6 +3,7 @@
 description = Korean peninsula hazard model 
 calculation_mode = classical
 random_seed = 23
+postproc_func = disagg_by_rel_sources
 
 [geometry]
