+++ openquake/qa_tests_data/logictree/case_05/job_bis.ini
@@ -3,6 +3,7 @@
 description = Reducing source specific logic tree, sampling
 calculation_mode = classical
 random_seed = 23
+postproc_func = disagg_by_rel_sources
 
 [geometry]
