+++ openquake/qa_tests_data/logictree/case_07/job.ini
@@ -5,6 +5,7 @@ calculation_mode = classical
 random_seed = 1066
 disagg_by_src = true
 use_rates = true
+postproc_func = disagg_by_rel_sources
 
 [geometry]
