+++ openquake/qa_tests_data/logictree/case_12/job.ini
@@ -3,6 +3,7 @@
 description = 2018 PSHA model - North Africa
 calculation_mode = classical
 random_seed = 19
+postproc_func = disagg_by_rel_sources
 
 [geometry]
