+++ openquake/qa_tests_data/logictree/case_19/job.ini
@@ -6,6 +6,7 @@ random_seed = 23
 disagg_by_src = true
 use_rates = true
 concurrent_tasks = 1
+postproc_func = disagg_by_rel_sources
 
 [geometry]
 sites = 16.2657 39.3002
